requirements.txt
flask
pandas
rapidfuzz
pyyaml
qrcode[pil]